Issue # 84
This will solve the GUI skeleton for peer-to-peer chat capabilities.  
The two files I had to change were MainMenu.fxml and
MenuController.java.  The MainMenu.fxml is what controlled the creation
of the sidebar and for the top bar.  I had to add an extra button on the
sidebar that would enable the use of the chat functionality of the
raiderplanner.  The second thing that I did was I edited the
MenuController.java file which is where I built the GUI for the project.
I created multiple methods that would help create the GUI for the chat
capabilities.  I also created getters and setters for the username and
for the hostname for future code implementation for the java peer to
peer chat.  I also createad my own imports and removed the ones that
were not being used by the program.