Added Error logging to the UIManager class
Added the option for reporting an error without logging it 
and just displaying it to the user.

Added the option for displaying an error to the user with a
custom log message.

Added the option for displaying an error to the user with a
log of the stack trace from a thrown exception.