Closes Issue #159
I removed the video class in model package. My reasoning for doing this
is becuase it is a dead class. It is only wasting space and I know that
the original intention was to use it for The requirement class to be
extended for specific cases, such as a video that needed to be watched,
or a book that needed to be read, with data specific for those cases.

This I felt could still be implemeted but for the time being since it is
not being used there was no sense in keeping the class. I liked the
idea for future developemet using videos which still could be used.
fixes #159