Adds limited implementation of add assignment button
This commit adds a limited implementation of the add assignment button to the assignment view pane. The button opens the add assignment UI, but currently the only fields that are added to the new assignment object are the assignment name and weighting fields. After closing the UI, the new assignment is displayed in the assignment view.