Completed add assignment button functionality
Added the coursework and exam assignment types, as well as all of the inputs that are required for each. The Assignments page now correctly displays the new assignments, and when the new assignments are clicked on, the details page now correctly displays and functions. Also added the set by field to the add assignments pop-up window.