GUI password implementation, email adaptation to include password structure, and password security strength
add labels for password and textbox for password inside account controller.java,
to allow users to create new accounts while using their session on the program.
add new email variable and structure to allow for this password to be sent to user's
personal email in account controller.java.
add conditionals as parameters for password strength as alerts asking for security strength,
and a check for pasword input at allow in account controller.java
modify wrightstudyprofile.java, person.java, and hudfile.java to allow for password implimenaation
modify createaccount.fxml to allow for sstyling of password lable and textbox for the entire
user GUI.

This fixes issue #190